makes port-mapper a library
